This is v0.2.8 ALPHA 13-Jun-2014:
Added a second pan/tilt unit mounted next to the gripper, to mount a Rasperry Pi camera module.

This camera is able to "see" what the gripper is near, for better positioning, as well as other
	things in other directions, and will also support advanced image processing.

Added defines in the header file and modified the initPanTilt() routine to initialize both pan/tilt
	units using the new setPanTiltHome() routine.

Added the setPanTiltHome() routine to initialize a single pan/tilt unit, called from initPanTilt().